--- Emby.Kodi.SyncQueue/Entities/DistinctRec.cs
@@ -1,7 +1,0 @@
-﻿namespace Emby.Kodi.SyncQueue.Entities
-{
-    public class DistinctRec
-    {
-        public string Id;
-    }
-}
