--- Emby.Kodi.SyncQueue/Entities/ServerTimeInfo.cs
@@ -1,14 +1,0 @@
-﻿namespace Emby.Kodi.SyncQueue.Entities
-{
-    public class ServerTimeInfo
-    {
-        public string ServerDateTime { get; set; }
-        public string RetentionDateTime { get; set; }
-
-        public ServerTimeInfo()
-        {
-            ServerDateTime = "";
-            RetentionDateTime = "";
-        }
-    }
-}
