--- Emby.Kodi.SyncQueue/Entities/UserJson.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Emby.Kodi.SyncQueue.Entities
-{
-    public class UserJson
-    {
-        public string Id { get; set; }
-        public string JsonData { get; set; }
-    }
-}
