+++ Jellyfin.Plugin.KodiSyncQueue/Configuration/configPage.html
