+++ Jellyfin.Plugin.KodiSyncQueue/Entities/DistinctRec.cs
@@ -1,0 +1,7 @@
+﻿namespace Jellyfin.Plugin.KodiSyncQueue.Entities
+{
+    public class DistinctRec
+    {
+        public string Id;
+    }
+}
