+++ Jellyfin.Plugin.KodiSyncQueue/Entities/LibFolder.cs
@@ -1,0 +1,11 @@
+﻿using MediaBrowser.Controller.Entities;
+using System;
+
+namespace Jellyfin.Plugin.KodiSyncQueue.Entities
+{
+    public class LibFolder
+    {
+        public Guid Id { get; set; }
+        public long SyncApiModified { get; set; }
+    }
+}
