+++ Jellyfin.Plugin.KodiSyncQueue/Entities/LibItem.cs
@@ -1,0 +1,19 @@
+﻿using MediaBrowser.Controller.Entities;
+using System;
+
+namespace Jellyfin.Plugin.KodiSyncQueue.Entities
+{
+    public class LibItem
+    {
+        public Guid Id { get; set; }
+        public long SyncApiModified { get; set; }
+        public int ItemType { get; set; }
+
+        // 0 = Movies
+        // 1 = TVShows
+        // 2 = Music
+        // 3 = Music Videos
+        // 4 = BoxSets
+
+    }
+}
