+++ Jellyfin.Plugin.KodiSyncQueue/Entities/PluginSettings.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Jellyfin.Plugin.KodiSyncQueue.Entities
+{
+    public class PluginSettings
+    {
+        public int RetentionDays { get; set; }
+        public bool IsEnabled { get; set; }
+        public bool TrackMovies { get; set; }
+        public bool TrackTVShows { get; set; }
+        public bool TrackMusic { get; set; }
+        public bool TrackMusicVideos { get; set; }
+        public bool TrackBoxSets { get; set; }
+    }
+}
