+++ Jellyfin.Plugin.KodiSyncQueue/Entities/ServerTimeInfo.cs
@@ -1,0 +1,14 @@
+﻿namespace Jellyfin.Plugin.KodiSyncQueue.Entities
+{
+    public class ServerTimeInfo
+    {
+        public string ServerDateTime { get; set; }
+        public string RetentionDateTime { get; set; }
+
+        public ServerTimeInfo()
+        {
+            ServerDateTime = "";
+            RetentionDateTime = "";
+        }
+    }
+}
