+++ Jellyfin.Plugin.KodiSyncQueue/Entities/UserJson.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Jellyfin.Plugin.KodiSyncQueue.Entities
+{
+    public class UserJson
+    {
+        public string Id { get; set; }
+        public string JsonData { get; set; }
+    }
+}
