+++ Jellyfin.Plugin.KodiSyncQueue/NanoApi/Entities/DbHeader.cs
