+++ Jellyfin.Plugin.KodiSyncQueue/NanoApi/Entities/Foo.cs
