+++ Jellyfin.Plugin.KodiSyncQueue/NanoApi/Entities/FooHelper.cs
