+++ Jellyfin.Plugin.KodiSyncQueue/NanoApi/Impersonate.cs
