+++ Jellyfin.Plugin.KodiSyncQueue/NanoApi/JsonFile.cs
